requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
autopep8==1.6.0
bcrypt==3.2.2
cffi==1.15.0
click==8.1.2
cryptography==37.0.2
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.75.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
Mako==1.2.0
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.17.1
toml==0.10.2
typing_extensions==4.2.0
uvicorn==0.17.6
zipp==3.8.0
